Refactor FuelService and TechnologyService to use BaseService architecture
