+++ src/services/utils/ServiceInitializer.ts
@@ -7,8 +7,8 @@ import { ServiceManager } from './ServiceManager';
 import { DataService } from '../core/DataService';
 import { RecipeService } from '../core/RecipeService';
 import { UserProgressService } from '../state/UserProgressService';
-import { TechnologyService } from '../TechnologyService';
-import { FuelService } from '../FuelService';
+import { TechnologyService } from '../game/TechnologyService';
+import { RefactoredFuelService as FuelService } from '../game/RefactoredFuelService';
 import { PowerService } from '../PowerService';
 import { StorageService } from '../StorageService';
 import { GameStorageService } from '../GameStorageService';
